requirements/prod.txt
djangae==0.9.6
django>=1.8,<1.10
git+https://github.com/mozilla/django-csp.git#egg=djangocsp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
six

# Uncomment these to use mapreduce or djangae.contrib.uniquetool (which uses mapreduce)
# These are forks of Google's libraries, which contain various fixes and patches
#git+https://github.com/potatolondon/appengine-pipelines.git
#git+https://github.com/potatolondon/potato-mapreduce.git#egg=mapreduce


djangorestframework==3.5.3
django-rest-swagger==0.3.5
pytz==2016.10
django-webpack-loader==0.4.1
git+https://github.com/kasajei/pyndlsearch.git#egg=pyndlsearch
requests_toolbelt
python-amazon-simple-product-api